# contracts/openapi-before.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/observabilidad/eventos": {
      "post": {
        "operationId": "TelemetryController_ingest",
        "summary": "Ingesta pública limitada de eventos de navegador",
        "description": "Telemetría no verificada, sin identidad de usuario. Máximo 20 eventos, 2 KiB por payload, 48 KiB por batch y 20 batches/minuto por IP. No acepta datos arbitrarios.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowserBatchDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Batch persistido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryAcceptedDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO, categoría, tipo, tamaño o payload inválido."
          },
          "429": {
            "description": "Límite de ingesta excedido."
          }
        },
        "tags": [
          "Observabilidad"
        ]
      }
    },
    "/api/v1/admin/observabilidad/stream": {
      "get": {
        "operationId": "AdminTelemetryController_stream",
        "summary": "Stream SSE de snapshots persistidos; JWT ADMIN en Authorization",
        "description": "Evento snapshot con JSON tras persistencia; reconciliacion cada 15 segundos. Reconectar cada 55 segundos para revalidar JWT. Sin tokens en URL.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "text/event-stream; event: snapshot; data: TelemetrySnapshotDto.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "403": {
            "description": "Solo ADMIN."
          }
        },
        "tags": [
          "Observabilidad ADMIN"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/admin/observabilidad/eventos": {
      "get": {
        "operationId": "AdminTelemetryController_recent",
        "summary": "Últimos 100 eventos de navegador persistidos y resumen de la muestra",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Eventos sanitizados y métricas reales de la muestra.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetrySnapshotDto"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "403": {
            "description": "Solo ADMIN."
          }
        },
        "tags": [
          "Observabilidad ADMIN"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/admin/observabilidad/resumen": {
      "get": {
        "operationId": "AdminTelemetryController_summary",
        "summary": "Resumen de los últimos 100 eventos de navegador",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Conteos y promedio de carga si existen al menos dos navegaciones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetrySummaryDto"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "403": {
            "description": "Solo ADMIN."
          }
        },
        "tags": [
          "Observabilidad ADMIN"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Booking Prototipo API",
    "description": "API activa del módulo de atracciones. Errores: 400 (validación), 401 (JWT), 403 (rol o propiedad), 404 (recurso) y 409 (disponibilidad, pago o idempotencia). Las rutas administrativas requieren Bearer JWT y rol ADMIN.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "BrowserEventDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "PERFORMANCE",
              "ERROR",
              "RESOURCE_ERROR",
              "INTERACTION",
              "VISIBILITY",
              "VIEWPORT",
              "CONNECTION",
              "CAPABILITY",
              "HTTP",
              "DOMAIN"
            ]
          },
          "type": {
            "type": "string",
            "description": "Tipo permitido para la categoría.",
            "maxLength": 50
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "route": {
            "type": "string",
            "maxLength": 200
          },
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador aleatorio en memoria, sin identidad de usuario."
          },
          "payload": {
            "type": "object",
            "description": "Solo campos técnicos permitidos; plano, máximo 2 KiB."
          }
        },
        "required": [
          "category",
          "type",
          "timestamp",
          "route",
          "sessionId",
          "payload"
        ]
      },
      "BrowserBatchDto": {
        "type": "object",
        "properties": {
          "events": {
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrowserEventDto"
            }
          }
        },
        "required": [
          "events"
        ]
      },
      "TelemetryAcceptedDto": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "accepted"
        ]
      },
      "PersistedBrowserEventDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "PERFORMANCE",
              "ERROR",
              "RESOURCE_ERROR",
              "INTERACTION",
              "VISIBILITY",
              "VIEWPORT",
              "CONNECTION",
              "CAPABILITY",
              "HTTP",
              "DOMAIN"
            ]
          },
          "type": {
            "type": "string",
            "description": "Tipo permitido para la categoría.",
            "maxLength": 50
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "route": {
            "type": "string",
            "maxLength": 200
          },
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador aleatorio en memoria, sin identidad de usuario."
          },
          "payload": {
            "type": "object",
            "description": "Solo campos técnicos permitidos; plano, máximo 2 KiB."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha del servidor; timestamp es reportado por el navegador."
          }
        },
        "required": [
          "category",
          "type",
          "timestamp",
          "route",
          "sessionId",
          "payload",
          "id",
          "receivedAt"
        ]
      },
      "TelemetrySummaryDto": {
        "type": "object",
        "properties": {
          "sampleSize": {
            "type": "number"
          },
          "sampleLimit": {
            "type": "number",
            "example": 100
          },
          "counts": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            },
            "description": "Conteos de eventos por categoría dentro de la muestra."
          },
          "navigationSamples": {
            "type": "number"
          },
          "averageLoadMs": {
            "type": "number",
            "nullable": true,
            "description": "Media de loadMs; null si hay menos de dos navegaciones completas."
          }
        },
        "required": [
          "sampleSize",
          "sampleLimit",
          "counts",
          "navigationSamples",
          "averageLoadMs"
        ]
      },
      "TelemetrySnapshotDto": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersistedBrowserEventDto"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/TelemetrySummaryDto"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "events",
          "summary",
          "generatedAt"
        ]
      },
...
    }
  }
}